move content info definition in specific files
