+++ src/app/modules/group/pages/group-by-id/group-by-id.component.ts
@@ -3,7 +3,8 @@ import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 import { FetchError, Fetching, isReady, Ready } from 'src/app/shared/helpers/state';
-import { CurrentContentService, GroupInfo, isGroupInfo } from 'src/app/shared/services/current-content.service';
+import { GroupInfo, isGroupInfo } from 'src/app/shared/models/content/group-info';
+import { CurrentContentService } from 'src/app/shared/services/current-content.service';
 import { ModeAction, ModeService } from 'src/app/shared/services/mode.service';
 import { Group } from '../../http-services/get-group-by-id.service';
 import { GroupDataSource } from '../../services/group-datasource.service';
