+++ src/app/shared/models/content/content-info.ts
@@ -1,0 +1,15 @@
+import { ContentBreadcrumb } from './content-breadcrumb';
+import { ContentRoute } from '../../routing/content-route';
+
+export interface ContentInfo {
+  type: string
+  breadcrumbs?: ContentBreadcrumb,
+  editing?: boolean, // undefined: not allowed, otherwise: whether the current page is current being edited
+  title?: string, // page title
+  route?: ContentRoute,
+  details?: any
+}
+
+export interface RoutedContentInfo extends ContentInfo {
+  route: ContentRoute
+}
